requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyyaml==6.0.1
gitpython==3.1.40
requests==2.31.0
python-dateutil==2.8.2
